Verson 1.0.250412

+++ src/lesson1/Trade.java
@@ -3,11 +3,13 @@ package src.lesson1;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class Trade {
     final long timestamp;
-    final double price;
-    final double amount;
+    final BigDecimal price;
+    final BigDecimal amount;
     final TradeType type;
 }
